LPFE-69: Moved week-menu.service to models and rename to menu
