--- src/app/components/menu/menu.service.ts
@@ -1,7 +1,0 @@
-import {IProduct} from '../../models/product';
-
-export interface IMenu {
-  id: number;
-  date: string;
-  products: IProduct[];
-}
